indexer/requirements.txt
psycopg2>=2.9.3
psycopg2>=2.8
python-dateutil>=2.5
tqdm==4.62.3
requests==2.26.0
peewee==3.14.8
pydantic
tqdm
dnspython
